Step 1 of huge update

+++ src/me/Man_cub/Buddies/world/lighting/BuddiesCuboidLightBuffer.java
@@ -4,19 +4,18 @@ import org.spout.api.lighting.Modifiable;
 import org.spout.api.util.cuboid.AlignedCuboidNibbleLightBuffer;
 
 public class BuddiesCuboidLightBuffer extends AlignedCuboidNibbleLightBuffer {
-	
 	protected BuddiesCuboidLightBuffer(BuddiesCuboidLightBuffer buffer) {
 		super(buffer);
 	}
-	
-	protected BuddiesCuboidLightBuffer(Modifiable holder, int id, int baseX, int baseY, int baseZ, int sizeX, int sizeY, int sizeZ) {
+
+	protected BuddiesCuboidLightBuffer(Modifiable holder, short id, int baseX, int baseY, int baseZ, int sizeX, int sizeY, int sizeZ) {
 		super(holder, id, baseX, baseY, baseZ, sizeX, sizeY, sizeZ);
 	}
 
-	protected BuddiesCuboidLightBuffer(Modifiable holder, int id, int baseX, int baseY, int baseZ, int sizeX, int sizeY, int sizeZ, byte[] data) {
+	protected BuddiesCuboidLightBuffer(Modifiable holder, short id, int baseX, int baseY, int baseZ, int sizeX, int sizeY, int sizeZ, byte[] data) {
 		super(holder, id, baseX, baseY, baseZ, sizeX, sizeY, sizeZ, data);
 	}
-	
+
 	@Override
 	public BuddiesCuboidLightBuffer copy() {
 		return new BuddiesCuboidLightBuffer(this);
